extract sources to own file
